Refactor: NaverWrapper 위치 변경 (repository -> service)
